finished the code, added details to the ending, making sure that every ending does end, and not leave an infinite loop.
Also added hints to the riddles that provide retry, maked the spelling of some of the words make sense. Still on trouble with the grammer of the story. Added the .toLowercase(), so everything that the user inputted will be in lowercase.

Updated the process of JavaDoc, have of the JavaDoc and design document are finished.